Fixed all the compiler warnings about unsigned long <-> int conversions.
● It seems there is real confusion in the code in the usage of `size_t`, `unsigned long` & `int`.

● I have fixed these issues by changing the types of the variables involved and in some cases minor changes to their usage.

● I hope this is not an indication of the quality of this code and I'm not wasting my time fixing this stuff 🤔

+++ src/generics/filesystem_handlers/file_path_decomposition.c
@@ -1,19 +1,19 @@
 /*
  The MIT License (MIT)
- 
+
  Copyright (c) 2016 Pomfort GmbH
  https://github.com/pomfort/mhl-tool
- 
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
- 
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -30,7 +30,7 @@
 #ifdef WIN
 #include <Windows.h>
 #include <direct.h>
-#endif 
+#endif
 
 #include <facade_info/error_codes.h>
 #include <generics/filesystem_handlers/public_interface.h>
@@ -68,23 +68,23 @@ void make_wpath_os_specific(wchar_t* wpath)
 }
 
 /**
- * @returns: 
- *   if path contains root path prefix ("C:", "d:\"), 
+ * @returns:
+ *   if path contains root path prefix ("C:", "d:\"),
  *   then returns size of path prefix
  *   if path does not contains root path prefix, then returns 0
  */
 unsigned char
 is_root_wpass_prefix(const wchar_t* wpath, size_t path_sz)
 {
-    wchar_t* os_specific_wpath = 0;    
+    wchar_t* os_specific_wpath = 0;
     const wchar_t* wpath_ptr = 0;
     unsigned char path_prefix_sz = 0;
-    
+
     if (wpath == 0 || path_sz < 2)
     {
         return 0;
     }
-    
+
     os_specific_wpath = mhlosi_wstrndup(wpath, path_sz);
     if (os_specific_wpath == 0)
     {
@@ -111,16 +111,16 @@ is_root_wpass_prefix(const wchar_t* wpath, size_t path_sz)
         free(os_specific_wpath);
         return path_prefix_sz;
     }
-    
+
     if (wpath_ptr[1] != L':')
     {
         free(os_specific_wpath);
         return path_prefix_sz;
     }
-    
+
     path_prefix_sz += 2;
     wpath_ptr += 2;
-    
+
     if (*wpath_ptr == L'\0')
     {
         free(os_specific_wpath);
@@ -129,9 +129,9 @@ is_root_wpass_prefix(const wchar_t* wpath, size_t path_sz)
 
     if (*wpath_ptr == WPATH_SEPARATOR)
     {
-      path_prefix_sz += 1;        
+      path_prefix_sz += 1;
     }
-    
+
     free(os_specific_wpath);
     return path_prefix_sz;
 }
@@ -143,18 +143,18 @@ are_wpaths_items_equal(const wchar_t* wpil, const wchar_t* wpir)
     {
         return 0;
     }
-    
+
     while (*wpil != L'\0' && *wpir != L'\0')
     {
         if (toupper(*wpil) != toupper(*wpir))
         {
             return 0;
         }
-        
+
         ++wpil;
         ++wpir;
-    } 
-    
+    }
+
     return *wpil != *wpir ? 0 : 1;
 }
 
@@ -170,8 +170,8 @@ void make_wpath_os_specific(wchar_t* wpath)
 }
 
 /**
- * @returns: 
- *   if path contains root path prefix ("/"), 
+ * @returns:
+ *   if path contains root path prefix ("/"),
  *   then returns size of path prefix, it is always 1
  *   if path does not contain root path prefix, then returns 0
  */
@@ -182,7 +182,7 @@ is_root_wpass_prefix(const wchar_t* wpath, size_t wpath_sz)
     {
         return 0;
     }
-    
+
     return wpath[0] == L'/' ? 1 : 0;
 }
 
@@ -193,34 +193,34 @@ are_wpaths_items_equal(const wchar_t* wpil, const wchar_t* wpir)
     {
         return 0;
     }
-    
+
     return wcscmp(wpil, wpir) == 0 ? 1 : 0;
 }
 
 #endif
 
-int 
+int
 add_sz_witem_to_fs_wpath(
-  const wchar_t* wnm, 
-  size_t wnm_sz, 
+  const wchar_t* wnm,
+  size_t wnm_sz,
   st_fs_wpath* p_wpath)
 {
     wchar_t**  new_witems;
     size_t witems_cnt;
-    
+
     if (wnm == 0 || wnm_sz == 0 || p_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     witems_cnt = p_wpath->items_cnt;
     new_witems = (wchar_t**) realloc(p_wpath->items, (witems_cnt + 1) * sizeof(wchar_t*));
     if (new_witems == 0)
     {
         return ERRCODE_OUT_OF_MEM;
-    } 
+    }
     p_wpath->items = new_witems;
-    
+
     p_wpath->items[witems_cnt] = (wchar_t*) calloc(wnm_sz + 1, sizeof(wchar_t));
     if (p_wpath->items[witems_cnt] == 0)
     {
@@ -229,42 +229,42 @@ add_sz_witem_to_fs_wpath(
     // We already have L'\0' at the end of string due to calloc
     wcsncpy(p_wpath->items[witems_cnt], wnm, wnm_sz);
     p_wpath->items_cnt = witems_cnt+1;
-    
-    return 0;  
+
+    return 0;
 }
 
-int 
+int
 add_witem_to_fs_wpath(const wchar_t* wnm, st_fs_wpath* p_wpath)
 {
     size_t sz;
-    
+
     if (wnm == 0 || p_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     sz = wcslen(wnm);
     if (sz == 0)
     {
         return 0;
     }
-    
+
     return add_sz_witem_to_fs_wpath(wnm, sz, p_wpath);
 }
 
-int 
+int
 remove_last_witem_from_fs_wpath(st_fs_wpath* p_wpath)
 {
     if (p_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     if (p_wpath->items_cnt == 0)
     {
         return ERRCODE_WRONG_FILE_LOCATION;
     }
-    
+
     free(p_wpath->items[p_wpath->items_cnt-1]);
     --p_wpath->items_cnt;
     return 0;
@@ -277,12 +277,12 @@ int init_fs_wpath(const wchar_t* wpath, st_fs_wpath* p_fs_wpath)
     unsigned char root_prefix_len;
     const wchar_t* wptr;
     size_t wsz;
-    
+
     if (p_fs_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     memset(p_fs_wpath, 0, sizeof(st_fs_wpath) / sizeof(char));
     wsz = wcslen(wpath);
     root_prefix_len = is_root_wpass_prefix(wpath, wsz);
@@ -290,21 +290,21 @@ int init_fs_wpath(const wchar_t* wpath, st_fs_wpath* p_fs_wpath)
     {
       p_fs_wpath->is_absolute = 1;
 
-      // if length of root prefix more than 1 wide character or 
-      //    last character in the root prefix is not parth separartor, 
+      // if length of root prefix more than 1 wide character or
+      //    last character in the root prefix is not parth separartor,
       // then it will be added to st_fs_wpath
       // Only windows root paths ("C:\" or "\\?\C:" are fit in this category
-      // If last character in the path prefix is filepasth separator, 
+      // If last character in the path prefix is filepasth separator,
       // it will be skipped.
-      // Later when path string will be constructed from content of 
+      // Later when path string will be constructed from content of
       // st_fs_wpath structure, skipped filepath separator, will be
       // added authomatically during filepath string construction
       if (wpath[root_prefix_len - 1] != WPATH_SEPARATOR || root_prefix_len > 1)
       {
-        res = 
+        res =
           add_sz_witem_to_fs_wpath(
-            wpath, 
-            wpath[root_prefix_len - 1] != WPATH_SEPARATOR ? root_prefix_len : root_prefix_len - 1, 
+            wpath,
+            wpath[root_prefix_len - 1] != WPATH_SEPARATOR ? root_prefix_len : root_prefix_len - 1,
             p_fs_wpath);
 
         if (res != 0)
@@ -316,7 +316,7 @@ int init_fs_wpath(const wchar_t* wpath, st_fs_wpath* p_fs_wpath)
 
       wpath += root_prefix_len;
     }
-    
+
     wptr = wcschr(wpath, WPATH_SEPARATOR);
     while (wptr)
     {
@@ -333,7 +333,7 @@ int init_fs_wpath(const wchar_t* wpath, st_fs_wpath* p_fs_wpath)
         wpath += wsz + 1;
         wptr = wcschr(wpath, WPATH_SEPARATOR);
     }
-    
+
     res = 0;
     if (*wpath != L'\0')
     {
@@ -343,7 +343,7 @@ int init_fs_wpath(const wchar_t* wpath, st_fs_wpath* p_fs_wpath)
           free_fs_wpath(p_fs_wpath);
         }
     }
-    
+
     return res;
 }
 
@@ -354,7 +354,7 @@ void free_fs_wpath(st_fs_wpath* p_fs_wpath)
   {
     return;
   }
-    
+
   for (i = 0; i < p_fs_wpath->items_cnt; ++i)
   {
     free(p_fs_wpath->items[i]);
@@ -364,14 +364,14 @@ void free_fs_wpath(st_fs_wpath* p_fs_wpath)
 }
 
 /*
- * Add path to destination, 
+ * Add path to destination,
  * added path will be normalized during adding process.
  */
 int add_normalized(st_fs_wpath* p_src_wpath, st_fs_wpath* p_dst_wpath)
 {
     int res;
     size_t i;
-    
+
     for (i = 0; i < p_src_wpath->items_cnt; ++i)
     {
         if (p_src_wpath->items[i] == 0 ||
@@ -380,7 +380,7 @@ int add_normalized(st_fs_wpath* p_src_wpath, st_fs_wpath* p_dst_wpath)
         {
             continue;
         }
-        
+
         if (wcscmp(p_src_wpath->items[i], L"..") == 0)
         {
             res = remove_last_witem_from_fs_wpath(p_dst_wpath);
@@ -389,8 +389,8 @@ int add_normalized(st_fs_wpath* p_src_wpath, st_fs_wpath* p_dst_wpath)
                 return res;
             }
         }
-        else 
-        {     
+        else
+        {
             res = add_witem_to_fs_wpath(p_src_wpath->items[i], p_dst_wpath);
             if (res != 0)
             {
@@ -398,57 +398,57 @@ int add_normalized(st_fs_wpath* p_src_wpath, st_fs_wpath* p_dst_wpath)
             }
         }
     }
-    
+
     return 0;
 }
 
 int convert_fs_wpath_to_absolute(
-  st_fs_wpath* p_base_wdir, 
+  st_fs_wpath* p_base_wdir,
   st_fs_wpath* p_fs_wpath)
 {
     int res;
     st_fs_wpath merged_fs_wpath;
-    
+
     if (p_fs_wpath && p_fs_wpath->is_absolute)
     {
         // target path is already absolute
         return 0;
     }
-    
+
     if (p_base_wdir == 0 &&  p_fs_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     if (p_base_wdir->is_absolute == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     // Create and fill temporary merged_fs_path
     memset(&merged_fs_wpath, 0, sizeof(st_fs_wpath) / sizeof(char));
-    
+
     res = add_normalized(p_base_wdir, &merged_fs_wpath);
     if (res != 0)
     {
         free_fs_wpath(&merged_fs_wpath);
         return res;
     }
-    
+
     res = add_normalized(p_fs_wpath, &merged_fs_wpath);
     if (res != 0)
     {
         free_fs_wpath(&merged_fs_wpath);
         return res;
     }
-    
+
     // move merged_fs_path into fs_path
     free_fs_wpath(p_fs_wpath);
     p_fs_wpath->is_absolute = 1;
     p_fs_wpath->is_normalized = 1;
     p_fs_wpath->items = merged_fs_wpath.items;
     p_fs_wpath->items_cnt = merged_fs_wpath.items_cnt;
-    
+
     return 0;
 }
 
@@ -457,26 +457,26 @@ int normalize_fs_wpath(st_fs_wpath* p_fs_wpath)
     unsigned char is_absolute;
     int res;
     st_fs_wpath merged_fs_wpath;
-    
+
     if (p_fs_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     if (p_fs_wpath->is_normalized)
     {
         return 0;
     }
-    
+
     // Create and fill temporary merged_fs_path
     memset(&merged_fs_wpath, 0, sizeof(st_fs_wpath) / sizeof(char));
-    
+
     res = add_normalized(p_fs_wpath, &merged_fs_wpath);
     if (res != 0)
     {
         return res;
     }
-    
+
     // swap fs_path objects
     is_absolute = p_fs_wpath->is_absolute;
     free_fs_wpath(p_fs_wpath);
@@ -484,7 +484,7 @@ int normalize_fs_wpath(st_fs_wpath* p_fs_wpath)
     p_fs_wpath->is_normalized = 1;
     p_fs_wpath->items = merged_fs_wpath.items;
     p_fs_wpath->items_cnt = merged_fs_wpath.items_cnt;
-    
+
     return 0;
 }
 
@@ -502,21 +502,21 @@ is_nested_fs_wpath(st_fs_wpath* p_fs_wbase, st_fs_wpath* p_fs_wcheck)
     {
         return 0;
     }
-    
+
     if (p_fs_wcheck->items_cnt < p_fs_wbase->items_cnt)
     {
         return 0;
     }
-    
+
     for (i = 0; i < p_fs_wbase->items_cnt; ++i)
     {
         if (!are_wpaths_items_equal(
                 p_fs_wbase->items[i], p_fs_wcheck->items[i]))
         {
             return 0;
-        }   
+        }
     }
-    
+
     return 1;
 }
 
@@ -529,48 +529,48 @@ fs_wpath_to_wstring(st_fs_wpath* p_rel_wpath, wchar_t** p_file_rel_wpath)
   size_t wln;
   wchar_t* wpointer;
   wchar_t** witems = p_rel_wpath->items;
-  
+
   if (p_file_rel_wpath == NULL)
   {
     //fprintf(stderr, "Buffer for filename is not empty.\n");
     return ERRCODE_INTERNAL_ERROR;
   }
-  
+
   if (p_rel_wpath->items_cnt == 0)
   {
     //fprintf(stderr, "Empty relative path to file.\n");
     return ERRCODE_INTERNAL_ERROR;
   }
-  
+
   if (p_rel_wpath->is_absolute)
   {
     wbuf_len = 1;
   }
-  
+
   for (i=0; i < p_rel_wpath->items_cnt; ++i)
   {
     wln = wcslen(witems[i]);
     if (wln == 0)
     {
       //fprintf(
-      //    stderr, 
+      //    stderr,
       //    "Unknown internal error: empty directory or file name "
       //    "in path.\n");
       return ERRCODE_INTERNAL_ERROR;
     }
-    wbuf_len += wln + 1; // add 1 symbol for PATH_SEPARATOR or 
+    wbuf_len += wln + 1; // add 1 symbol for PATH_SEPARATOR or
     // '\0' at the end
   }
-  
+
   *p_file_rel_wpath = (wchar_t*)calloc(wbuf_len, sizeof(wchar_t));
   if (*p_file_rel_wpath == NULL)
   {
     //fprintf(stderr, "Out of memory.\n");
     return ERRCODE_OUT_OF_MEM;
   }
-  
-  
-  // Construct relative path to file as a one string 
+
+
+  // Construct relative path to file as a one string
   wpointer = *p_file_rel_wpath;
 
 #ifndef WIN
@@ -581,23 +581,23 @@ fs_wpath_to_wstring(st_fs_wpath* p_rel_wpath, wchar_t** p_file_rel_wpath)
     ++wpointer;
   }
 #endif
-  
+
   for (i=0; i < p_rel_wpath->items_cnt - 1; ++i)
   {
     wln = wcslen(witems[i]);
     if (wln)
     {
       wcsncpy(wpointer, witems[i], wln);
-      
+
       wpointer[wln] = WPATH_SEPARATOR;
       wpointer += wln + 1;
     }
   }
-  
+
   wln = wcslen(witems[i]);
   wcsncpy(wpointer, witems[i], wln);
   // We already have '\0' at the end of line due to calloc
-  
+
   return 0;
 }
 
@@ -605,23 +605,23 @@ fs_wpath_to_wstring(st_fs_wpath* p_rel_wpath, wchar_t** p_file_rel_wpath)
 // Paths p_base_path and p_file_path must be absolute
 int
 extract_relative_fs_wpath(
-  st_fs_wpath* p_base_wpath, 
-  st_fs_wpath* p_file_wpath, 
+  st_fs_wpath* p_base_wpath,
+  st_fs_wpath* p_file_wpath,
   st_fs_wpath* p_rel_wpath)
 {
     size_t i;
     int res;
-    
+
     if (p_base_wpath == 0 || p_file_wpath == 0 || p_rel_wpath == 0)
     {
         return ERRCODE_INTERNAL_ERROR;
     }
-    
+
     if (!is_nested_fs_wpath(p_base_wpath, p_file_wpath))
     {
         return ERRCODE_WRONG_FILE_LOCATION;
     }
-    
+
     for (i = p_base_wpath->items_cnt; i < p_file_wpath->items_cnt; ++i)
     {
         res = add_witem_to_fs_wpath(p_file_wpath->items[i], p_rel_wpath);
@@ -630,39 +630,39 @@ extract_relative_fs_wpath(
             return res;
         }
     }
-    
+
     return 0;
 }
 
 int
 extract_relative_wpath(
-  st_fs_wpath* p_base_wpath, 
+  st_fs_wpath* p_base_wpath,
   st_fs_wpath* p_file_wpath,
   wchar_t** p_file_rel_wpath)
 {
     int res;
     st_fs_wpath rel_wpath;
-    
+
     memset(&rel_wpath, 0, sizeof(rel_wpath) / sizeof(char));
-    
+
     res = extract_relative_fs_wpath(p_base_wpath, p_file_wpath, &rel_wpath);
     if (res != 0)
     {
       return res;
     }
-    
+
     res = fs_wpath_to_wstring(&rel_wpath, p_file_rel_wpath);
     if (res != 0)
     {
       return res;
     }
-    
+
     free_fs_wpath(&rel_wpath);
-    
+
     return 0;
 }
 
-/* Note! Caller is responsible for freeing allocated string 
+/* Note! Caller is responsible for freeing allocated string
  *
  * Create path string from first n items of st_fs_wpath.
  * String with path will be allocated and stored in p_wpath.
@@ -676,37 +676,37 @@ fs_wspath_first_nitems_to_str(
   size_t i;
   size_t wln;
   size_t dst_wpath_sz;
-  int real_n;
+  size_t real_n;
   wchar_t** witems;
   wchar_t* wpointer;
   unsigned char is_item_added = 0;
-  
+
   if (p_fs_wpath == 0 || p_dst_wpath == 0)
   {
     return ERRCODE_WRONG_ARGUMENTS;
   }
-  
+
   // calculate requires size of destination string
   real_n = p_fs_wpath->items_cnt > (size_t) first_n ? (size_t) first_n : p_fs_wpath->items_cnt;
   dst_wpath_sz = 0;
-  for (i = 0; i < (size_t) real_n; ++i)
+  for (i = 0; i < real_n; ++i)
   {
     dst_wpath_sz += wcslen(p_fs_wpath->items[i]);
   }
-  dst_wpath_sz += (real_n + 2); 
-  
+  dst_wpath_sz += (real_n + 2);
+
   // alloacte wstring
   *p_dst_wpath = calloc(dst_wpath_sz, sizeof(wchar_t));
   if (*p_dst_wpath)
   {
     return ERRCODE_OUT_OF_MEM;
   }
-  
+
   // construct answer
   witems = p_fs_wpath->items;
   wpointer = *p_dst_wpath;
-  
-#ifndef WIN  
+
+#ifndef WIN
   // In Linux and MacOSX root path inital slash is not in "items" list
   if (p_fs_wpath->is_absolute)
   {
@@ -714,9 +714,9 @@ fs_wspath_first_nitems_to_str(
     ++wpointer;
   }
 #endif
-  
-  for (i = 0; i < (size_t) real_n; ++i)
-  {    
+
+  for (i = 0; i < real_n; ++i)
+  {
     wln = wcslen(witems[i]);
     if (wln)
     {
@@ -726,17 +726,17 @@ fs_wspath_first_nitems_to_str(
       is_item_added = 1;
     }
   }
-  
+
   // remove last WPATH_SEPARATOR
   if (is_item_added)
   {
     *(wpointer - 1) = L'\0';
   }
-  
+
   return 0;
 }
 
-/* 
+/*
  * Create st_fs_wpath structure started from n-th item of p_fs_wpath.
  */
 int
@@ -744,10 +744,10 @@ fs_wspath_from_nitems_to_fs_wspath(
   st_fs_wpath* p_fs_wpath,
   int from_n,
   st_fs_wpath* p_rest_wpath)
-{ 
+{
   int res;
   size_t i;
-  
+
   if (p_fs_wpath == 0 || p_rest_wpath == 0)
   {
     return ERRCODE_WRONG_ARGUMENTS;
@@ -757,24 +757,23 @@ fs_wspath_from_nitems_to_fs_wspath(
   if ((size_t) from_n >= p_fs_wpath->items_cnt)
   {
     // nothing to copy from_n is beyond the border
-    return 0; 
+    return 0;
   }
-  
+
   for (i = from_n; i < p_fs_wpath->items_cnt; ++i)
   {
-    res = 
+    res =
       add_sz_witem_to_fs_wpath(
-        p_fs_wpath->items[i], 
+        p_fs_wpath->items[i],
         wcslen(p_fs_wpath->items[i]),
         p_rest_wpath);
-    
+
     if (res != 0)
     {
       free_fs_wpath(p_rest_wpath);
       return res;
     }
   }
-  
+
   return 0;
 }
-
